+++ docs/source/user_guide/changelog.rst
@@ -31,6 +31,7 @@ Unreleased
 - Remove parameter ``IDENTITY_TRANSFER`` setting it to ``True`` if a knowledge transfer folder is provided.
 - Fix bug in the :ref:`validator_reference` when saving session from the pop up dialog at closing the GUI.
 - Fix bug in the :ref:`validator_reference` when opening a session from the pop up dialog at starting the GUI.
+- Add identities to fragments when tracking single animal or single fragment videos. This allows to use idmatcher.ai afterwards.
 
 6.0.12
 ======